+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/Persons/PersonDetail/EditInduction/InductionChangeReason.cshtml
@@ -36,7 +36,7 @@
                     <govuk-radios-item value="@true">
                         Yes
                         <govuk-radios-item-conditional data-testid="additional-detail">
-                            <govuk-character-count asp-for="ChangeReasonDetail" label-class="govuk-label--m" max-length="InductionDefaults.DetailMaxCharacterCount" rows="InductionDefaults.DetailTextAreaMinimumRows" />
+                            <govuk-character-count asp-for="ChangeReasonDetail" label-class="govuk-label--m" max-length="FileUploadDefaults.DetailMaxCharacterCount" rows="FileUploadDefaults.DetailTextAreaMinimumRows" />
                         </govuk-radios-item-conditional>
                     </govuk-radios-item>
                     <govuk-radios-item value="@false">
