+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/RoutesToProfessionalStatus/EditRoute/CheckYourAnswers.cshtml.cs
@@ -2,10 +2,10 @@ using System.Reflection.Metadata.Ecma335;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using TeachingRecordSystem.Core;
 using TeachingRecordSystem.Core.DataStore.Postgres;
 using TeachingRecordSystem.Core.Services.Files;
 using TeachingRecordSystem.SupportUi.Infrastructure.Filters;
-using TeachingRecordSystem.SupportUi.Pages.Persons.PersonDetail.EditInduction;
 
 namespace TeachingRecordSystem.SupportUi.Pages.RoutesToProfessionalStatus.EditRoute;
 
@@ -19,7 +19,7 @@ public class CheckYourAnswersModel(
     public JourneyInstance<EditRouteState>? JourneyInstance { get; set; }
 
     public string? PersonName { get; set; }
-
+    public Guid PersonId { get; private set; }
     public QualificationType? QualificationType { get; set; }
     public Guid RouteToProfessionalStatusId { get; set; }
     public ProfessionalStatusStatus Status { get; set; }
@@ -48,6 +48,36 @@ public class CheckYourAnswersModel(
     public string[]? TrainingSubjects { get; set; }
     public string? RouteToProfessionalStatusName { get; set; }
 
+    public async Task OnGetAsync()
+    {
+        RouteToProfessionalStatusName = (await referenceDataCache.GetRouteToProfessionalStatusByIdAsync(RouteToProfessionalStatusId))?.Name!;
+        ExemptionReason = InductionExemptionReasonId is not null ? (await referenceDataCache.GetInductionExemptionReasonByIdAsync(InductionExemptionReasonId!.Value))?.Name : null;
+        TrainingProvider = TrainingProviderId is not null ? (await referenceDataCache.GetTrainingProviderByIdAsync(TrainingProviderId!.Value))?.Name : null;
+        TrainingCountry = TrainingCountryId is not null ? (await referenceDataCache.GetTrainingCountryByIdAsync(TrainingCountryId))?.Name : null;
+        TrainingSubjects = TrainingSubjectIds is not null ?
+            TrainingSubjectIds
+                .Join((await referenceDataCache.GetTrainingSubjectsAsync()), id => id, subject => subject.TrainingSubjectId, (_, subject) => subject.Name)
+                .OrderByDescending(name => name)
+                .ToArray() : null;
+    }
+
+    public async Task<IActionResult> OnPostAsync()
+    {
+        // save the changes
+
+        // broadcast event
+
+        await JourneyInstance!.CompleteAsync();
+
+        return Redirect(linkGenerator.PersonQualifications(PersonId));
+    }
+
+    public async Task<IActionResult> OnPostCancelAsync()
+    {
+        await JourneyInstance!.DeleteAsync();
+        return Redirect(linkGenerator.PersonQualifications(PersonId));
+    }
+
     public override async Task OnPageHandlerExecutionAsync(PageHandlerExecutingContext context, PageHandlerExecutionDelegate next)
     {
         JourneyInstance!.State.EnsureInitialized(context.HttpContext.GetCurrentProfessionalStatusFeature());
@@ -59,7 +89,8 @@ public class CheckYourAnswersModel(
 
         var personInfo = context.HttpContext.GetCurrentPersonFeature();
         PersonName = personInfo.Name;
-        
+        PersonId = personInfo.PersonId;
+
         ChangeReason = JourneyInstance!.State.ChangeReason;
         ChangeReasonDetail = JourneyInstance!.State.ChangeReasonDetail!;
 
@@ -77,23 +108,13 @@ public class CheckYourAnswersModel(
         TrainingProviderId = JourneyInstance!.State.TrainingProviderId;
         InductionExemptionReasonId = JourneyInstance!.State.InductionExemptionReasonId;
 
-        // CML TODO move to get?
-        RouteToProfessionalStatusName = (await referenceDataCache.GetRouteToProfessionalStatusByIdAsync(RouteToProfessionalStatusId))?.Name!;
-        ExemptionReason = InductionExemptionReasonId is not null ? (await referenceDataCache.GetInductionExemptionReasonByIdAsync(InductionExemptionReasonId!.Value))?.Name : null;
-        TrainingProvider = TrainingProviderId is not null ? (await referenceDataCache.GetTrainingProviderByIdAsync(TrainingProviderId!.Value))?.Name : null;
-        TrainingCountry = TrainingCountryId is not null ? (await referenceDataCache.GetTrainingCountryByIdAsync(TrainingCountryId))?.Name : null;
-        TrainingSubjects = TrainingSubjectIds is not null ?
-            TrainingSubjectIds
-                .Join((await referenceDataCache.GetTrainingSubjectsAsync()), id => id, subject => subject.TrainingSubjectId, (_, subject) => subject.Name)
-                .OrderByDescending(name => name)
-                .ToArray() : null;
         await next();
     }
 
     public async Task<string?> GetEvidenceFileUrlAsync()
     {
         return ChangeReasonDetail.EvidenceFileId is not null ?
-            await fileService.GetFileUrlAsync(ChangeReasonDetail.EvidenceFileId!.Value, InductionDefaults.FileUrlExpiry) : // CML TODO - move to a general defaults file
+            await fileService.GetFileUrlAsync(ChangeReasonDetail.EvidenceFileId!.Value, FileUploadDefaults.FileUrlExpiry) :
             null;
     }
 }
